Abstract out path to data
'Data' relative to the framework is primarily post and page files as well as aggregrate data structures such as tags, navigation lists, and the global site hash. These structures are pseudo generated, acting only as stubs so the templates have data to preview correctly. In the real site generation scenario, a server language is needed to build the data-structures by traversing the post/pages.