Better to have the ruby logic as utility functions.
This is my take on functional programming but I may just be making stuff up. Essentially we want the user to have the freedom to use utilities piece-meal, so we call each thing we need one a time, or the user can comment it out if he wants. Who knows, but seems simpler to me - yell at me if this is terrible.