Switch back to mustache
Going to try to implement basic helpers in mustache because while Handlebars has more functionality it is not ported in very many languages. The entire reason I started with mustache is with the idea of being able to process the templates in the user's language of choice. Let's get to it