Add basic helper functionality to mustache
This is a proof of concept. The idea is to catch 'not found' mustache expressions that are also preceded with a '?'. This identifies the expression as a call to a helper. The helper is responsible for querying the global object hash based on a 'context' which can be passed as part of the helper's name deliminated by a colon ':'. We say 'query' to mean basic data manipulation to deliver the proper object back into the normal mustache 'renderSection' calling context. In other words our helpers are merely data-structure manipulations, that pass the manipulated data-structures back into the calling context as if it was a normal call.