Add stub framework for mustache parsing in ruby
This is entirely proof of concept. Finally got the thing working. The main concept is to overload the find method in order to 'catch' tagnames with a '?' and pass them to my custom view methods. The tagname format has to change since ruby mustache actually evalulates one context at a time when working with dot notation. The js mustache version evaluates the entire tagname as one context. Clearer explanations to follow.